requirements.txt
# 車載端依賴（Raspberry Pi）
opencv-python==4.8.1.78
numpy==1.24.3
pyserial==3.5
RPi.GPIO==0.7.1
requests==2.31.0
pynmea2==1.19.0
geopy==2.4.0
python-dotenv==1.0.0
flask==3.0.0

